chore: rename ContactPage
